feat: introduce session dynamics
The TCSVTrialsBase.AssignParameters can't be used to determine
dinamic parameters, because it is called before the session starts.

Session dynamics are trial parameters that may change during the
session. They may be deduced from the session configuration file or
they may be created during the session.

For now, only a function to set negative comparisons is implemented.
The function determines which S- words are used in the trial, based
on experiment specific rules (for now hardcoded ones).

The position of stimuli on screen is also dynamic, but it is implemented
in separate units (grids). Positions and negative stimuli are recorded
in the session's "{XXX}.data" file, in real time, trial by trial.

The order of trials inside a block is also dynamic, but it is
implemented in separate units (shuffler). However, the order of trials
inside the configuration file must be preserved, because it is
determined by trial unique IDs. Hence, order is determined using a mask
for unique IDS. A new order (mask) is generated at the beginning of each
block.

Shuffling requires trials with ReferenceNames and unique IDs. IDs are
necessary to mask conversion. When enough ReferenceNames are available
inside a block, the shuffler will distribute trials in groups of unique
ReferenceNames, avoiding more than two consecutive trials with the same
ReferenceName inside the block. The result order is recorded in the
session's "{XXX}.data" file, in real time, trial by trial.

TODO: refact the code to improve readability about dinamic parameters
and features.